+++ services/information_intelligence/__init__.py
@@ -3,5 +3,6 @@
 from .openapi import canonicalize_openapi, diff_openapi, structural_contract
 from .legal import canonicalize_legal, canonicalize_legal_pdf, diff_legal
 from .manual_evidence import SELLER_MAIN_NOTICES, daily_brief_preview
+from .news import canonicalize_news, diff_news
 
-__all__ = ["canonicalize_openapi", "diff_openapi", "structural_contract", "canonicalize_legal", "canonicalize_legal_pdf", "diff_legal", "SELLER_MAIN_NOTICES", "daily_brief_preview"]
+__all__ = ["canonicalize_openapi", "diff_openapi", "structural_contract", "canonicalize_legal", "canonicalize_legal_pdf", "diff_legal", "SELLER_MAIN_NOTICES", "daily_brief_preview", "canonicalize_news", "diff_news"]